change Federator file name to avoid overwritting on compiling
